% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMTloop.R
\name{doMTloop}
\alias{doMTloop}
\title{doMT: estimate mpt model using multiTree}
\usage{
doMTloop(fEqn, data, catlabels = NULL, restrictions = NULL, fOut = "",
  lambda = 0, numRep = 3, maxIter = 10000, convergence = 1e-10,
  CI = 95, bootstrapSamples = 0, fia = FALSE)
}
\arguments{
\item{fEqn}{eqn file location.}

\item{data}{either a string pointing to a mdt file or a N*M data matrix or a N*M data frame.}

\item{catlabels}{A string vector of category labels (matching the heading of the data file), only required for data matrix.}

\item{restrictions}{= either a string pointing to a restrictions file or a list containing restrictions e.g. list("a=b", "a=.5").}

\item{fOut}{filename where outputfile should be written.}

\item{lambda}{lambda value for PD^lamda used in minimization and fit assessment.}

\item{numRep}{how many times parameter estimation should be repeated (to avoid local minima).}

\item{maxIter}{maximum number of iterations.}

\item{convergence}{convergence criterion.}

\item{CI}{size of confidence interval in \%, e.g. 95 for 95\%}

\item{fia}{whether to compute the fisher information approximation of the minimum description length.}
}
\description{
Estimate an MPT model using multiTree, but using a loop for data (ensures that FIA is computed correctly for different N per tree)
}

