% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeR.R
\name{summary.multiTreeR}
\alias{summary.multiTreeR}
\title{summary writes summary given mt result object}
\usage{
\method{summary}{multiTreeR}(mtres)
}
\arguments{
\item{mtres}{the mt result object (\code{\link{doMT}})}
}
\description{
summary writes summary given mt result object
}

