% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeR-package.R
\docType{package}
\name{multiTreeR}
\alias{multiTreeR}
\alias{multiTreeR-package}
\title{multiTreeR: R interface for performing analyses of MPT models using multiTree.}
\description{
TO fit an MPT model using multiTree in R use \code{\link{doMT}}
}
\details{
multiTreeR calls java code, so the  Java Runtime Environment has to be installed, get it here: \url{http://www.java.com}.
For details on multiTree, see \url{http://psycho3.uni-mannheim.de/multitree}
}
\author{
Morten Moshagen
}
\references{
Moshagen, M. (2010). multiTree: A computer program for the analysis of multinomial processing tree models. Behavior Research Methods, 42, 42-54.
}

