% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.multiTreeR}
\alias{summary.multiTreeR}
\title{summary writes summary given mt result object}
\usage{
\method{summary}{multiTreeR}(mtres, digits = 4, ci = FALSE)
}
\arguments{
\item{mtres}{the mt result object (\code{\link{doMT}})}

\item{digits}{number digits to round}

\item{ci}{whether to show confidence intervals}
}
\description{
summary writes summary given mt result object
}

